sk/requirements.txt
azure-ai-projects 
azure-identity 
semantic-kernel[azure] 
aiohttp 
python-dotenv
streamlit